split object_cache into its own file
